Add files via upload
This is a build up onto the previous assignment. The scenario remains the same as for the previous: the simulation of a pay-to-stay car paid. In thisproject we bring it to a conclusion by adding in the missing element: payment.
To recap, the kind of car park in question has a ticket machine at the entrance and a cashier at the exit. A driver, on entering the car park receives a ticket stamped with the arrival time. (The arrival time is also recorded on the magnetic strip on the back.) On exit, the driver gives the ticket to the cashier, the duration of the stay is calculated and from that, how much must be paid. For this project, the focus is on developing a class of “tariff table” that can be used to determine the cost of a stay.

This is done by firstly developing a TimePeriod class of an object. A TimePeriod is a duration range. It has an inclusive lower bound, l, and an exclusive upper bound,
u. A time duration, d, falls within the range if l≤ d < u. 

Secondly the ParkingTariff class is developed. Each ParkingTariff object stores a TimePeriod and Money.

Thirdly, I now develop the TariffTable class. A TariffTable contains a collection of ParkingTariff objects.

Lastly I extended the CarParSim solution that you constructed for the previous project by:
• Adding code to create and populate a TariffTable object.
• Extending the code for the ‘depart’ command to include printing the tariff that applies to a stay of that duration.
• Adding code for an additional ‘tariffs’ command: When the user enters the ‘tariffs’ command, a list of parking tariffs will be printed.